Use unsigned integer instead of

+++ database/migrations/2018_04_11_073030_create_donations_table.php
@@ -19,7 +19,7 @@ class CreateDonationsTable extends Migration
 
             $table->string('title');
             $table->longText('description');
-            $table->double('target')->nullable();
+            $table->unsignedInteger('target')->nullable();
             $table->text('location')->nullable();
             $table->text('video_url')->nullable();
             $table->timestamp('end_at')->nullable();
